update fork logic

+++ app/fork_test.go
@@ -16,7 +16,7 @@ import (
 func TestFork(t *testing.T) {
 	realio := Setup(false, nil)
 
-	ctx := realio.BaseApp.NewContext(false, tmproto.Header{Height: int64(ForkHeight)})
+	ctx := realio.BaseApp.NewContext(false, tmproto.Header{Height: ForkHeight})
 	stakingKeeper := realio.StakingKeeper
 
 	timeKey := time.Date(2024, 4, 1, 1, 1, 1, 1, time.UTC)
@@ -25,7 +25,7 @@ func TestFork(t *testing.T) {
 		DelegatorAddress: "test_del_1",
 		ValidatorAddress: "test_val_1",
 		Entries: []stakingtypes.UnbondingDelegationEntry{
-			stakingtypes.NewUnbondingDelegationEntry(int64(ForkHeight), timeKey, math.OneInt()),
+			stakingtypes.NewUnbondingDelegationEntry(ForkHeight, timeKey, math.OneInt()),
 		},
 	}
 
@@ -37,7 +37,7 @@ func TestFork(t *testing.T) {
 		ValidatorSrcAddress: "test_val_1",
 		ValidatorDstAddress: "test_val_2",
 		Entries: []stakingtypes.RedelegationEntry{
-			stakingtypes.NewRedelegationEntry(int64(ForkHeight), timeKey, math.OneInt(), sdk.OneDec()),
+			stakingtypes.NewRedelegationEntry(ForkHeight, timeKey, math.OneInt(), sdk.OneDec()),
 		},
 	}
 	stakingKeeper.InsertRedelegationQueue(ctx, duplicativeRedelegation, timeKey)
@@ -46,7 +46,7 @@ func TestFork(t *testing.T) {
 
 	duplicativeVal := stakingtypes.Validator{
 		OperatorAddress: "test_op",
-		UnbondingHeight: int64(ForkHeight),
+		UnbondingHeight: ForkHeight,
 		UnbondingTime:   timeKey,
 	}
 
@@ -72,7 +72,7 @@ func TestFork(t *testing.T) {
 	require.Equal(t, triplets[0].ValidatorDstAddress, duplicativeRedelegation.ValidatorDstAddress)
 	require.Equal(t, triplets[0].ValidatorSrcAddress, duplicativeRedelegation.ValidatorSrcAddress)
 
-	vals := stakingKeeper.GetUnbondingValidators(ctx, timeKey, int64(ForkHeight))
+	vals := stakingKeeper.GetUnbondingValidators(ctx, timeKey, ForkHeight)
 	require.Equal(t, vals[0], duplicativeVal.OperatorAddress)
 
 }
